BepsApi/Backend/API/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
cryptography==41.0.3
flask-jwt-extended==4.7.1
python-dotenv==1.0.1
requests>=2.31.0
urllib3<2 
concurrent-log-handler==0.9.25
six>=1.17.0
pandas==2.2.3
websockets